third_party/blink/web_tests/platform/generic/fast/dom/icon-size-property-expected.txt

Tests for sizes attribute of the link element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS icon.sizes.length is 2
PASS icon.sizes.value is "23x47 19x42"
PASS icon.sizes.length is 3
PASS icon.sizes.value is "10x10 20x20 30x30"
PASS icon.sizes.length is 0
PASS icon.sizes.value is ""
PASS icon.sizes.length is 0
PASS icon.sizes.value is ""
PASS successfullyParsed is true

TEST COMPLETE
